Finally figured out resizable divs
Yasssss, I finally figured out how to make two resizable divs for the
website!!!!! In addition, I made a very basic design for the resource
button because I can't really think of one right now. Small changes to
the z-indexes of the dropdowns and the shading div when you open the
sidebar.