Made some changes to the resizing calculations
Changed the resizing calculations so that the final result would be a
percentage of the width of the page rather than just pixels. The problem
with pixels was that if the user shrunk the browser width to a width
smaller than the size of the left div's pixel size, the site would try
to grow to fit the div.